Derive debug trait to the solution struct
Each field does implement Debug, T is a FloatT i believe and SolverStatus also implements it. During developement using good_lp i runt into a bug and was trying to print a solution but it was not possible since this, i would find this can help me and anyone who would also want to debug a solution.